Rename iconbar.c and iconbar.h to ibar to prevent confusion with tboxlib iconbar.h
